Rename to eliminate_agg_distinct
